--- src/main/java/roomescape/exception/advice/ErrorResponse.java
@@ -1,7 +1,0 @@
-package roomescape.exception.advice;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public record ErrorResponse(String code, String message) {
-}
