+++ src/main/java/roomescape/exception/handler/ErrorResponse.java
@@ -1,0 +1,4 @@
+package roomescape.exception.handler;
+
+public record ErrorResponse(String code, String message) {
+}
